use size_t for json strings

+++ src/mimo.h
@@ -14,7 +14,7 @@ class mimo_model_data
 public:
   /** query size of json encoding of model (including terminating 0).
       will be followed by a call to to_json() */
-  virtual int json_size () = 0;
+  virtual size_t json_size () = 0;
 
   /** output as json string into out
       Is preceeded by a call to json_size().
@@ -30,7 +30,7 @@ public:
       separate one from the main app. We can't even work on a
       std::stringstream passed by reference, since probably the
       reallocations also use the heap. */
-  virtual char *to_json (char *out, int size) throw() = 0;
+  virtual char *to_json (char *out, size_t size) throw() = 0;
 
   /** get model from json string */
   virtual int from_json (const char *json_string) = 0;
